### 2018-11-26 #267 @Bisaloo
lightr
`lightr` imports [UV-VIS reflectance/transmission/absorbance](https://en.wikipedia.org/wiki/Ultraviolet%E2%80%93visible_spectroscopy) proprietary file formats in R. It also allows the import of related metadata that are critical to ensure reproducibility but that are often discarded by other tools.

```
Package: lightr
Title: Read Spectrometric Data in R
Version: 0.0.0.9000
    person("Hugo", "Gruson", role = c("cre", "aut"),
    email = REDACTED,
    person("Rafael", "Maia", role = "aut",
    email = "REDACTED",
    person("Thomas", "White", role = "aut",
    email = "REDACTED",
    )
Description: Parse various UV-VIS reflectance/transmittance/absorbance spectra
    file formats to extract spectral data and metadata.
    pbmcapply,
    xml2
    covr,
License: GPL (>=2)
Language: en-GB
```

https://github.com/Bisaloo/lightr

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under **and why**? (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.):

    - data extraction, because the package parses multiple scientific data file formats
    - (reproducibility) because it also extracts metadata saved during the recording

People working with UV-VIS reflectance/transmittance/absorbance spectra, colour science (#colsci). People developing package to analyse spectral data with vision models (`lightr` has few dependencies). Even non-R users who do not own the proprietary software to convert the proprietary formats to `csv` files.

There is partial overlap with some other packages, as described in the [README](https://github.com/Bisaloo/lightr#similar-projects) but to my knowledge, none of them have the same aims as `lightr`.

- [ ] contains a vignette with examples of its essential functions and uses. **[[in progress](https://bisaloo.github.io/lightr/articles/)]**

Some function names do not meet rOpenSci criteria (e.g. `getspec()` and `getmetadata()`) but this is kept for backwards compatibility with the package `pavo`, from which `lightr` originated. Synonyms have been added (`get_spec()` and `get_metadata()`) but it is at the moment unlikely that the old names will be deprecated.

----

As mentioned in the template, some pieces still need polishing, mainly the vignettes but before going for one last push, I'd like to know if you would be interested.
