### 2018-11-27 90585b1
Add more about uniqueness of lightr
