### 2018-11-28 8556c6d
Add examples to parsers
