+++ R/parse_generic.R
@@ -10,6 +10,11 @@
 #'   * a dataframe with columns "wl", "dark", "white", "scope" in that order
 #'   * a list with metadata including
 #'
+#' @examples
+#' parse_generic(system.file("testdata", "spec.csv", package = "lightr"),
+#'               sep = ",")
+#' parse_generic(system.file("testdata", "CRAIC_export.txt", package = "lightr"))
+#'
 #' @export
 #'
 
