+++ R/parse_avantes_binary.R
@@ -6,6 +6,10 @@
 #'
 #' @importFrom stats setNames
 #'
+#' @examples
+#' parse_trm(system.file("testdata", "avantes_trans.TRM", package = "lightr"))
+#' parse_roh(system.file("testdata", "avantes_reflect.ROH", package = "lightr"))
+#'
 #' @export
 #'
 parse_trm <- function(filename) {
