### 2018-11-29 0105868
Compare to known output in tests
