### 2018-11-29 426eea8
Add dispatch tests
