### 2018-11-29 365de68
Add loop to spec2csv
