### 2018-11-29 712eb5f
Add minimal info to batch import vignette
