### 2018-11-29 359092e
Add huxtable to suggests
