### 2018-11-29 b8bd921
Finish custom import vignette
