+++ DESCRIPTION
@@ -19,7 +19,6 @@ Imports:
     xml2
 Suggests:
     covr,
-    huxtable,
     knitr,
     rmarkdown,
     spelling,
