### 2018-11-29 3998d9b
Remove extension check
