### 2018-11-29 a381202
Drop unused arg from docs
