+++ R/spec2csv.R
@@ -1,7 +1,5 @@
 #' Convert spectral data files to csv files
 #'
-#' @param filename name of the file
-#'
 #' @param overwrite logical. Should the function overwrite existing files with
 #' the same name? (defaults to `FALSE`).
 #'
