### 2018-11-30 cc717bd
Set up goodpractice with travis
