+++ R/getmetadata.R
@@ -108,7 +108,9 @@ get_metadata <- function(where = getwd(), ext = "ProcSpec", sep = NULL,
     paste(c("white", "dark", "sample"), "boxcar", sep = "_")
   )
 
-  res[, c(6,7,8,9,10,11,12,13,14)] <- sapply(res[, c(6,7,8,9,10,11,12,13,14)], as.numeric)
+  res[, c(6,7,8,9,10,11,12,13,14)] <- vapply(res[, c(6,7,8,9,10,11,12,13,14)],
+                                             as.numeric,
+                                             numeric(nrow(res)))
 
   return(res)
 }
