+++ .travis.yml
@@ -12,6 +12,7 @@ sudo: false
 cache: packages
 
 after_success:
+  - Rscript -e 'goodpractice::gp()'
   - Rscript -e 'covr::codecov()'
 
 before_deploy: Rscript -e 'install.packages("pkgdown")'
