### 2018-11-30 2907125
Set stringsAsFactors to FALSE
