### 2018-11-30 c6296c1
Edit regex to make it to the message error
