### 2018-11-30 0f70e76
Add tests for spec2csv
