### 2018-11-30 bc97976
Fix extension in tests
