### 2018-11-30 9ec9b9e
Fix typos
