### 2018-12-01 ccbd937
Fix tests
