### 2018-12-01 bf3c0d3
Recreate procspec folder when copying
