### 2018-12-03 c75d74b
Use recursive copy
