### 2018-12-03 dabe006
Do not copy subfolder content
