### 2018-12-03 09fcd64
Ouput numeric df
