### 2018-12-03 15fd21c
Add better defaults for dispatch parser
