### 2018-12-03 5526743
Make sure that all data outputs are df
