### 2018-12-04 7e8e3e5
Add more dispatch tests
