+++ tests/testthat/test_dispatch.R
@@ -6,6 +6,42 @@ test_that("Fallback", {
     parse_generic(test.file("spec.csv"), sep = ","),
     dispatch_parser(test.file("spec.csv"), sep = ",")
   )
+
+  expect_equal(
+    parse_jdx(test.file("OceanOptics.jdx")),
+    dispatch_parser(test.file("OceanOptics.jdx"))
+  )
+
+  expect_equal(
+    parse_jaz(test.file("jazspec.jaz")),
+    dispatch_parser(test.file("jazspec.jaz"))
+  )
+
+  expect_equal(
+    parse_jazirrad(test.file("irrad.JazIrrad")),
+    dispatch_parser(test.file("irrad.JazIrrad"))
+  )
+
+  expect_equal(
+    parse_roh(test.file("avantes_reflect.ROH")),
+    dispatch_parser(test.file("avantes_reflect.ROH"))
+  )
+
+  expect_equal(
+    parse_trm(test.file("avantes_trans.TRM")),
+    dispatch_parser(test.file("avantes_trans.TRM"))
+  )
+
+  expect_equal(
+    parse_ttt(test.file("avantes_export.ttt")),
+    dispatch_parser(test.file("avantes_export.ttt"))
+  )
+
+  expect_equal(
+    parse_trt(test.file("avantes_export2.trt")),
+    dispatch_parser(test.file("avantes_export2.trt"))
+  )
+
 })
 
 test_that("Similar output for all parsers", {
