### 2018-12-04 0353103
Remove jdx from 1st test wave
