### 2018-12-04 aad071d
Disable warn/error tests for spec2csv It looks like an error with expression(), where files cannot be found.
