### 2018-12-04 c99fcfb
Restore convert warn/error tests
