### 2018-12-04 308e059
Revert "Restore convert warn/error tests"
This reverts commit c99fcfbefd66e4903c88be0652a637bd43673f60.
