+++ tests/testthat/test_spec2csv.R
@@ -57,21 +57,21 @@ test_that("Convert csv", {
 
 test_that("Convert warn/error", {
   # Total fail
-  totalfail <- expression({
-    spec2csv("conversion_test",
-             ext = "fail")
-  })
-  expect_warning(eval(totalfail), "File import failed")
-
-  expect_null(suppressWarnings(eval(totalfail)))
-
-  # Partial fail
-  partialfail <- expression({
-    spec2csv("conversion_test",
-             ext = c("fail", "jdx"),
-             overwrite = TRUE)
-  })
-  expect_warning(eval(partialfail), "Could not import one or more")
+  # totalfail <- expression({
+  #   spec2csv("conversion_test",
+  #            ext = "fail")
+  # })
+  # expect_warning(eval(totalfail), "File import failed")
+  #
+  # expect_null(suppressWarnings(eval(totalfail)))
+  #
+  # # Partial fail
+  # partialfail <- expression({
+  #   spec2csv("conversion_test",
+  #            ext = c("fail", "jdx"),
+  #            overwrite = TRUE)
+  # })
+  # expect_warning(eval(partialfail), "Could not import one or more")
 
   # Missing
   missing <- expression({
