### 2018-12-04 d6d5836
Add wl range test to get_spec
