### 2018-12-05 69f11bf
Always return spectral data as numeric
