### 2018-12-05 b1982cc
Fix awkward path settings in vignette
