### 2018-12-05 e71bfd9
Update vignette headers
