+++ paper.Rmd
@@ -3,13 +3,21 @@ title: "paper"
 tags:
 author: 
   - name: Hugo Gruson
-    orcid:
+    orcid: "0000-0001-8493-9450"
     affiliation: 1
   - name: Thomas White
+    orcid: "0000-0002-4094-1476"
+    affiliation: 2
   - name: Rafael Maia
+    orcid: "0000-0002-7563-9795"
+    affiliation: 3
 affiliations:
-  - name: Centre d'Écologie Fonctionnelle et Évolutive
+  - name: Centre d'Écologie Fonctionnelle et Évolutive, Montpellier
     index: 1
+  - name: School of Life and Environmental Sciences, University of Sidney
+    index: 2
+  - name: Columbia University
+    index: 3
 date: "November 28, 2018"
 output: md_document
 ---
@@ -24,9 +32,9 @@ Spectrometers record the amount of captured photons at different wavelengths
 (usually between 300-700 nm for colour sicence). However, there is no standard 
 file format for spectrometry data and different scientific instrumentation 
 companies use wildly different formats to store spectral data. Vendors 
-proprietary software sometimes have an option but convert those formats instead 
+proprietary software sometimes have an option to convert those formats into
 human readable files such as `.csv` but such softwares are often expensive and
-they discard most metadata in the process. However, those metadata are critical
+they discard most metadata in the process. Yet, those metadata are critical
 to ensure reproducibility of the measurements, and ultimately of the scientific
 findings (@White2015).
 
@@ -38,6 +46,40 @@ solution to read proprietary spectra file formats on all systems.
 
 `lightr` started as a fork from the popular R package `pavo`, which provides
 a large suite of colour analysis tools (@Maia2013, @White2018).
+
+# Package design
+
+`lightr` has been designed to provide two levels on the complexity / 
+customability trade-off:
+
+* Spectral data and metadata for each file format are extracted using 
+specialized parsers. Parsers are also aliased with many different names so that
+users can often use `parse_$extension()` where `$extension$` is the file 
+extension of the file to parse. For convenience, we also provide a generic 
+fallback, named `parse_generic()` that works for many "simple" formats, often
+derived from `csv` or `tsv`. Specialized parsers should usually be preferred to
+`parse_generic()` because `parse_generic()` is not able to parse metadata.
+
+* Because spectrometers store each measurement in a separate file, the number of
+files for a single study can quickly increase. To ensure easy and efficient 
+processing of those files, `lightr` also provides three high-levels functions
+that can recursively find files and process them with a parallelized loop using
+the `parallel` R package.
+
+# Recommended workflow
+
+As mentioned earlier, proprietary spectrometry softwares can also export 
+spectral data into a human-readable format (usually a kind of tabulation
+separated values, or `tsv`, with a complex header). `lightr` can read files 
+generated by this export step. We however
+**do not recommend you use the software built-in export function**,
+because it will apply possibly unwanted transformation to your data
+(interpolation and  subsetting) and may discard important metadata.
+
+Instead, we recommend you keep the files in the proprietary format (such as
+Avantes `ABS`, `ROH` and `TRM`, or OceanOptics `ProcSpec` and `jdx`) and that
+you use `lightr` to convert them into your preferred file format (such as 
+`csv`). 
 
 # Acknowledgements
 
