### 2018-12-06 6f126e4
Install goodpractice on travis
