### 2018-12-06 a62c0b8
Install pkgdown on travis with standard instructions
