### 2018-12-07 cdf269f
Do not catch warnings
