+++ inst/lightr.bib
@@ -1,0 +1,79 @@
+
+@article{Cuthill1999_plumagereflectanceobjective,
+  langid = {english},
+  title = {Plumage Reflectance and the Objective Assessment of Avian Sexual Dichromatism},
+  volume = {153},
+  doi = {10.1086/303160},
+  number = {2},
+  journaltitle = {The American Naturalist},
+  date = {1999-09-09},
+  pages = {183-200},
+  author = {Cuthill, Innes C. and Bennett, Andrew T. D. and Partridge, J. C. and Maier, E. J.}
+}
+
+@article{Bennett1994_sexualselectionmismeasure,
+  langid = {english},
+  title = {Sexual Selection and the Mismeasure of Color},
+  volume = {144},
+  doi = {10.1086/285711},
+  number = {5},
+  journaltitle = {The American Naturalist},
+  date = {1994-11-01},
+  pages = {848-860},
+  author = {Bennett, Andrew T. D. and Cuthill, Innes C. and Norris, K. J.}
+}
+
+@article{Maia2013_pavopackageanalysis,
+  langid = {english},
+  title = {Pavo: An {{R}} Package for the Analysis, Visualization and Organization of Spectral Data},
+  volume = {4},
+  doi = {10.1111/2041-210X.12069},
+  shorttitle = {Pavo},
+  number = {10},
+  journaltitle = {Methods in Ecology and Evolution},
+  shortjournal = {Methods Ecol Evol},
+  date = {2013-10-01},
+  pages = {906-913},
+  author = {Maia, Rafael and Eliason, Chad M. and Bitton, Pierre-Paul and Doucet, Stéphanie M. and Shawkey, Matthew D.}
+}
+
+@article{White2015_signaldesigncourtship,
+  langid = {english},
+  title = {Signal Design and Courtship Presentation Coincide for Highly Biased Delivery of an Iridescent Butterfly Mating Signal},
+  volume = {69},
+  doi = {10.1111/evo.12551},
+  shorttitle = {Signal Design and Courtship Presentation Coincide for Highly Biased Delivery of an Iridescent Butterfly Mating Signal},
+  number = {1},
+  journaltitle = {Evolution},
+  date = {2015-01},
+  pages = {14-25},
+  author = {White, Thomas E. and Zeil, Jochen and Kemp, Darrell J.}
+}
+
+@article{Eaton2005_humanvisionfails,
+  langid = {english},
+  title = {Human Vision Fails to Distinguish Widespread Sexual Dichromatism among Sexually “Monochromatic” Birds},
+  volume = {102},
+  doi = {10.1073/pnas.0501891102},
+  number = {31},
+  journaltitle = {Proceedings of the National Academy of Sciences},
+  shortjournal = {PNAS},
+  date = {2005-08-02},
+  pages = {10942-10946},
+  author = {Eaton, Muir D.},
+  eprinttype = {pmid},
+  eprint = {16033870}
+}
+
+@article{Maia2018_pavonewtools,
+  langid = {english},
+  title = {Pavo 2.0: New Tools for the Spectral and Spatial Analysis of Colour in {{R}}},
+  doi = {10.1101/427658},
+  shorttitle = {Pavo 2.0},
+  journaltitle = {bioRxiv},
+  date = {2018-09-26},
+  pages = {427658},
+  author = {Maia, Rafael and Gruson, Hugo and Endler, John A. and White, Thomas E.}
+}
+
+
