### 2018-12-10 #267 @karthik
### Editor checks:

---

#### Editor comments

Package looks to be in excellent shape and your test coverage is fantastic! Goodpractice had very little to say and the only notable suggestion is to drop spelling from your suggests. I'm looking for appropriate reviewers now but if you have any suggestions for independent reviewers, please let me know. 🙏

```
  ✖ fix this R CMD check ERROR: Package suggested but not
available: ‘spelling’ The suggested packages are required for a complete check. Checking can be attempted without them by setting the environment variable _R_CHECK_FORCE_SUGGESTS_ to a false value. See section ‘The DESCRIPTION file’ in the ‘Writing R Extensions’ manual.
```

---

Reviewers: TBD
Due date: TBD
