### 2018-12-11 #267 @Bisaloo
Thank you for your comment!

I don't really understand why `spelling` is problematic. It's used in many packages on CRAN without any issue (https://cran.r-project.org/package=spelling). Could you expand a bit more why I should remove it please?

Also, it is okay if I keep working on `master` until reviewers are found?
