### 2018-12-13 cd22a33
Add "processed" column in docs
