+++ man/parse_jdx.Rd
@@ -12,7 +12,8 @@ parse_jdx(filename)
 \value{
 A list of two elements:
 \itemize{
-\item a dataframe with columns "wl", "dark", "white", "scope" in that order
+\item a dataframe with columns "wl", "dark", "white", "scope" and "processed",
+in that order
 \item a list with metadata including
 }
 }
