### 2018-12-13 e96fb53
Remove synonyms (breaks pavo transparent compatibility)
