+++ man/get_metadata.Rd
@@ -1,15 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/getmetadata.R
+% Please edit documentation in R/get_metadata.R
 \name{get_metadata}
 \alias{get_metadata}
-\alias{getmetadata}
 \title{Extract metadata from spectra files}
 \usage{
 get_metadata(where = getwd(), ext = "ProcSpec", sep = NULL,
-  subdir = FALSE, subdir.names = FALSE, cores = getOption("mc.cores",
-  2L), ignore.case = TRUE)
-
-getmetadata(where = getwd(), ext = "ProcSpec", sep = NULL,
   subdir = FALSE, subdir.names = FALSE, cores = getOption("mc.cores",
   2L), ignore.case = TRUE)
 }
