+++ man/get_spec.Rd
@@ -1,15 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/getspec.R
+% Please edit documentation in R/get_spec.R
 \name{get_spec}
 \alias{get_spec}
-\alias{getspec}
 \title{Extract spectral data from spectra files}
 \usage{
 get_spec(where = getwd(), ext = "txt", lim = c(300, 700),
-  decimal = ".", sep = NULL, subdir = FALSE, subdir.names = FALSE,
-  cores = getOption("mc.cores", 2L), ignore.case = TRUE)
-
-getspec(where = getwd(), ext = "txt", lim = c(300, 700),
   decimal = ".", sep = NULL, subdir = FALSE, subdir.names = FALSE,
   cores = getOption("mc.cores", 2L), ignore.case = TRUE)
 }
