### 2018-12-13 a5f5842
Use get_metadata in tests
