+++ tests/testthat/test_metadata.R
@@ -2,7 +2,7 @@ context("get_metadata")
 
 test_that("get_metadata all", {
 
-  res <- getmetadata(
+  res <- get_metadata(
     test.file(),
 	  ext = c("TRM", "ROH", "ttt", "trt", "jdx", "jaz", "JazIrrad")
   )
@@ -13,7 +13,7 @@ test_that("get_metadata all", {
 test_that("get_metadata recursive", {
 
   # Recursive
-  res <- getmetadata(test.file(),
+  res <- get_metadata(test.file(),
                      ext = "ProcSpec", subdir = TRUE)
   expect_equal_to_reference(res, "known_output/getmetadata_recursive.rds")
 })
@@ -21,7 +21,7 @@ test_that("get_metadata recursive", {
 test_that("get_metadata warn/error", {
   # Total fail
   totalfail <- expression({
-    getmetadata(test.file(),
+    get_metadata(test.file(),
                 ext = "fail")
   })
   expect_warning(eval(totalfail), "File import failed")
@@ -30,14 +30,14 @@ test_that("get_metadata warn/error", {
 
   # Partial fail
   partialfail <- expression({
-    getmetadata(test.file(),
+    get_metadata(test.file(),
                 ext = c("fail", "jdx"))
   })
   expect_warning(eval(partialfail), "Could not import one or more")
 
   # Missing
   missing <- expression({
-    getmetadata(ext = "missing")
+    get_metadata(ext = "missing")
   })
   expect_warning(eval(missing), "No files found")
 
