### 2018-12-13 ae9d8fd
Add more info about wl polynomes
