### 2018-12-13 e01f7f4
Add file type agnostic get_uninterp
