### 2018-12-13 fcbf164
Load lightr in vignette
