+++ DESCRIPTION
@@ -20,6 +20,7 @@ Imports:
 Suggests:
     covr,
     knitr,
+    pavo,
     rmarkdown,
     spelling,
     testthat
