### 2018-12-14 30cc300
Fix misc typos
