+++ .travis.yml
@@ -22,3 +22,5 @@ deploy:
   provider: script
   script: Rscript -e 'pkgdown::deploy_site_github()'
   skip_cleanup: true
+  on:
+    r: release
