### 2018-12-17 bac95f3
Only deploy on release version
