+++ .travis.yml
@@ -22,5 +22,4 @@ deploy:
   provider: script
   script: Rscript -e 'pkgdown::deploy_site_github()'
   skip_cleanup: true
-  on:
-    r: release
+  on: condition: "$TRAVIS_R_VERSION_STRING = release"
