### 2018-12-17 8733018
Remove covr from suggests
