### 2018-12-17 8640a36
Restart travis tests
