+++ .travis.yml
@@ -23,4 +23,5 @@ deploy:
   provider: script
   script: Rscript -e 'pkgdown::deploy_site_github()'
   skip_cleanup: true
-  on: condition: "$TRAVIS_R_VERSION_STRING = release"
+  on:
+    condition: "$TRAVIS_R_VERSION_STRING = release"
