+++ man/convert_tocsv.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/spec2csv.R
-\name{spec2csv}
-\alias{spec2csv}
+% Please edit documentation in R/convert_tocsv.R
+\name{convert_tocsv}
+\alias{convert_tocsv}
 \title{Convert spectral data files to csv files}
 \usage{
-spec2csv(where = getwd(), ext = "txt", decimal = ".", sep = NULL,
-  subdir = FALSE, cores = getOption("mc.cores", 2L),
+convert_tocsv(where = getwd(), ext = "txt", decimal = ".",
+  sep = NULL, subdir = FALSE, cores = getOption("mc.cores", 2L),
   ignore.case = TRUE, overwrite = FALSE)
 }
 \arguments{
