+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(convert_tocsv)
 export(get_metadata)
 export(get_spec)
 export(parse_abs)
@@ -12,7 +13,6 @@ export(parse_roh)
 export(parse_trm)
 export(parse_trt)
 export(parse_ttt)
-export(spec2csv)
 importFrom(pbmcapply,pbmclapply)
 importFrom(stats,approx)
 importFrom(stats,setNames)
