+++ _pkgdown.yml
@@ -9,7 +9,7 @@ reference:
     contents:
     - get_spec
     - get_metadata
-    - spec2csv
+    - convert_tocsv
   - title: Low-level parsers
     desc: These functions are destined to advanced users that want a more flexible way to import spectral data. They can only read a single file at a time and users are responsible for writing a loop if necessary.
     contents:
