### 2019-01-05 603a3b2
Update tests with convert_tocsv
