### 2019-01-05 4fdc36e
Remove last remaining spec2csv
