### 2019-01-05 6480c50
Update vignette titles to prevent duplicates
