### 2019-01-14 5081965
Improve approx speed and mem usage See https://bugs.r-project.org/bugzilla/show_bug.cgi?id=17515
