### 2019-01-15 7f19d5c
Restore macOS testing
