### 2019-01-15 3c52ab3
Install libgit on macOS
