+++ .travis.yml
@@ -14,6 +14,7 @@ r_packages:
   - covr
   - goodpractice
   - pkgdown
+brew_packages: libgit2
 
 after_success:
   - Rscript -e 'goodpractice::gp()'
