### 2019-01-15 ddbe4e3
Revert testing on macOS Broken and it probably won't be fixed: https://travis-ci.community/t/r-devel-el-capitan-signed-pkg-fails-to-download-when-testing-on-r-devel/672/5
