### 2019-01-23 b1945ae
Use expect_known_value
