### 2019-01-23 7d5e2b3
Clean up sorting stuff
