+++ R/get_metadata.R
@@ -44,10 +44,10 @@ get_metadata <- function(where = getwd(), ext = "ProcSpec", sep = NULL,
 
   extension <- paste0("\\.", ext, "$", collapse = "|")
 
-  file_names <- sort(list.files(where,
+  file_names <- list.files(where,
     pattern = extension, ignore.case = ignore.case,
     recursive = subdir, include.dirs = subdir
-  ))
+  )
 
   # This step is needed to ensure reproducibility between locales and platforms
   file_names <- sort(file_names, method = "radix")
