### 2019-01-23 402c6c3
Keep spectra names as character
