### 2019-01-24 61f6c12
Add minimal version to dependencies
