### 2019-01-24 #267 @Bisaloo
@karthik, thanks for the update :+1: .

Here are some people with a colour science background that code:

- Felipe Gawryszewski is the author of the [colourvision](https://cran.r-project.org/web/packages/colourvision/index.html) R package but I just had a look and he doesn't import spectra. The user has to provide a dataframe as far as I understand.
- Cassie Stoddard developed the [tetracolorspace](https://www.marycstoddard.com/software/) Matlab program but I don't know how much she uses R and as far as I understand, tetracolorspace only imports a small subset of spectra files (those that are "csv-like")
- [As mentioned in the source code](https://github.com/Bisaloo/lightr/blob/master/R/parse_avantes_binary.R#L16-L18), one function in `lightr` is an almost direct translation from a matlab script I found online. I don't know the author at all (I don't even know if that email is still working) but they should be very qualified to review `lightr` (assuming they have some R knowledge).

Beyond that, as I said, I don't think it is required to have a background in colour science to review `lightr`. It is not (as opposed to `colourvision` or `pavo`) a package to perform colour vision models. It focuses on data import and manipulation so I'm confident anyone with a solid general knowledge of R could provide a good feedback.
