### 2019-01-25 f882160
Set up release auto deploy
