+++ .travis.yml
@@ -1,5 +1,3 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
-
 language: R
 r:
   - oldrel
@@ -14,14 +12,20 @@ r_packages:
   - covr
   - goodpractice
   - pkgdown
-
 after_success:
   - Rscript -e 'goodpractice::gp()'
   - Rscript -e 'covr::codecov()'
-
 deploy:
-  provider: script
-  script: Rscript -e 'pkgdown::deploy_site_github()'
-  skip_cleanup: true
-  on:
-    condition: "$TRAVIS_R_VERSION_STRING = release"
+  - provider: script
+    script: Rscript -e 'pkgdown::deploy_site_github()'
+    skip_cleanup: true
+    on:
+      condition: "$TRAVIS_R_VERSION_STRING = release"
+  - provider: releases
+    api_key:
+      secure: uJsKCTuYpiFamNKUxMwy0MzHXWloq5dmJTyIoW7PUS//RREiWHDkoOXRK5FWhZzL+nbjTkmIh3sygAfW8kTBq+TJL0kaYNSrCSKJijFKqlh5ZBTyuV28nFYiT1Wx6YH7D+Bx74S0apvDnZ53Nn0ip9UZ8tOnutIYs+JLyqTbRYR8aRgp6uVQLlQmxSbQ1r1Q6kpuQ6eh2Aurc71urQGK9tFMzj6T0PA9sKt1VEkURUpfP0IX2qGBkfrlxjdz1m3h5uggO+FKWVo1CTUfpop2LcPdufF4nrfV70OVvpxoDoud+c2si+5lr/GrUHvqm2rRmc7d8TAh5FhvMD1m9B+YlQGLsYc/ExblC9S8OrSTCjMQZSSWOfw3PG/UkF6cZolnQmsv1mt6WJE62kOVnaby0lN/Hx2V4LLDbMmVGxyvOJwrpuPQECeNfp2M3IEHJUmeDB4k4EZmUpS3ohuohNvuKBt0J9+xK2uXEcGDaQx4Fm/REvK8JGJz6h1V0urPgn/xTzKnSFQF8zdJb1w67MATfF5XhDo8mOB5/jHkUrvnWqkV/6uqV2XezlgK1rZSyeupAiPwg/CGFir1JTxpEfKyeAHotuA0QhMSFw1aD0wsEakDoKuaNHqUKaO4q1l2GKTiMmR5U2f4epINMPI6Bwperewg7wJemnaM+ax7aWdFilA=
+    file_glob: true
+    file: lightr_*.tar.gz
+    skip_cleanup: true
+    on:
+      tags: true
