### 2019-01-25 b8ab930
Fix travis YAML syntax
