### 2019-01-25 0392117
Move test.file() function to main folder
