+++ tests/testthat.R
@@ -1,8 +1,4 @@
 library(testthat)
 library(lightr)
 
-test.file <- function(...) {
-  system.file("testdata", ..., package = "lightr")
-}
-
 test_check("lightr")
