+++ R/dev_utils.R
@@ -1,0 +1,3 @@
+test.file <- function(...) {
+  system.file("testdata", ..., package = "lightr")
+}
