### 2019-01-31 b4bb39a
Rename spec2csv in readme
