+++ README.md
@@ -1,6 +1,6 @@
 # `lightr`: import spectral data in R
 
-[![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.org/#wip)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![Travis build status](https://travis-ci.org/Bisaloo/lightr.svg?branch=master)](https://travis-ci.org/Bisaloo/lightr)
 [![Coverage status](https://codecov.io/gh/Bisaloo/lightr/branch/master/graph/badge.svg)](https://codecov.io/github/Bisaloo/lightr?branch=master)
 
