### 2019-02-01 ab7638d
Add tests on windows
