### 2019-02-02 f166fdd
Try to test on release only for macOS
