### 2019-02-02 db6e7aa
Fix travis syntax
