+++ .travis.yml
@@ -1,14 +1,14 @@
 language: R
 r:
   - oldrel
-  - release:
-    os:
-      - linux
-      - osx
+  - release
   - devel
 os:
   - linux
-#  - osx
+matrix:
+  include:
+    - r: release
+      os: osx
 sudo: false
 cache: packages
 r_packages:
