### 2019-02-02 0735546
Install libgit on macos
