+++ .travis.yml
@@ -9,6 +9,7 @@ matrix:
   include:
     - r: release
       os: osx
+      brew_packages: libgit2
 sudo: false
 cache: packages
 r_packages:
