### 2019-02-03 29a2f17
Rename functions in multicore test file
