+++ tests/testthat/test_multicores.R
@@ -4,11 +4,11 @@ test_that("Multicores", {
 
   skip_if(.Platform$OS.type != "windows")
 
-  expect_message(getspec(system.file("testdata", package = "lightr"),
+  expect_message(get_spec(system.file("testdata", package = "lightr"),
                          ext = "jdx", cores = 2),
                  '"cores" set to 1.')
 
-  expect_message(getmetadata(system.file("testdata", package = "lightr"),
+  expect_message(get_metadata(system.file("testdata", package = "lightr"),
                              ext = "jdx", cores = 2),
                  '"cores" set to 1.')
 
