### 2019-02-03 05a1c51
Try to set up stages in travis
