### 2019-02-03 9d26a54
Remove AND
