### 2019-02-03 5d73782
Fix travis syntax
