### 2019-02-04 30a0b35
Remove redundant test
