+++ .travis.yml
@@ -1,7 +1,6 @@
 language: R
 r:
   - oldrel
-  - release
   - devel
 os:
   - linux
