### 2019-02-04 #267 @Bisaloo
Here is the profile for a potential reviewer IMO. Maybe you will know someone who fits this description:

Someone who is familiar with file manipulation (`list.files()`, `path()`, `tempdir()`, `file_ext()`) and import (`unzip()`, `file()`, `readLines()`) in R, including binary files (`readBin()`, `intToUtf8()`) and XML files (via the xml2 package). Some metadata is imported through string manipulation (`grep()`, `gsub()`, `strsplit()`) as well but the regex patterns generally remain simple.

I'll try to add a vignette to describe the structure / internal organisation of the package at some point, which should greatly help for the review.
