### 2019-02-05 afd657d
Fix parser for OO files without header
