### 2019-02-05 9c97615
Stricter tests for parsers output uniformity
