+++ tests/testthat/test_dispatch.R
@@ -55,6 +55,8 @@ test_that("Similar output for all parsers", {
     expect_length(res, 2)
     expect_is(res[[1]], "data.frame")
     expect_true(all(apply(res[[1]], 2, is.numeric)))
+    expect_named(res[[1]], c("wl", "dark", "white", "scope", "processed"))
+    expect_length(res[[2]], 13)
   })
 
 })
