### 2019-02-05 3952d41
Make sure avantes converted files always output 5 columns as well
