### 2019-02-08 57f95a1
Simplify failures handling
