+++ R/get_spec.R
@@ -86,13 +86,13 @@ get_spec <- function(where = getwd(), ext = "txt", lim = c(300, 700),
              error = function(e) NULL
     ), mc.cores = cores)
 
-  if (any(unlist(lapply(tmp, is.null)))) {
-    whichfailed <- which(unlist(lapply(tmp, is.null)))
-    if (length(whichfailed) == nb_files) {
-      warning("File import failed.\n",
-              "Check input files and function arguments.", call. = FALSE)
-      return()
-    }
+  whichfailed <- which(vapply(tmp, is.null, logical(1)))
+
+  if (length(whichfailed) == nb_files) {
+    warning("File import failed.\n",
+            "Check input files and function arguments.", call. = FALSE)
+    return(NULL)
+  } else if (length(whichfailed) > 0) {
 
     warning("Could not import one or more files:\n",
       paste0(files[whichfailed], "\n"),
@@ -101,7 +101,6 @@ get_spec <- function(where = getwd(), ext = "txt", lim = c(300, 700),
 
     specnames <- specnames[-whichfailed]
   }
-
 
   tmp <- do.call(cbind, tmp)
 
