### 2019-06-05 #267 @Bisaloo
Thank you for your review @jeroen! You bring up very valid points but I'm not yet quite sure if/how I can address all of them. I'll need a couple of days to think about it. Here is a first attempt at answering some of your concerns. I will also perform some changes on the packages following your recommendations whenever I get a moment (hopefully next week).

> but there is little documentation on where to find the specification of these formats (if one exists)

To my knowledge, the only format that has been defined as a standard are `jdx` files. I do say it in the docs but I could try to think of a more prominent place where to put the reference:

https://github.com/Bisaloo/lightr/blob/b06c2b3c5512ff798c7ce82933c003bbdb4422bc/R/parse_jdx.R#L7-L9

> The parsing unit tests verify that the output contains data and metadata, but do not verify the data itself.

You are correct that I don't compare the output to known results from an outside source but I however implemented regression tests to compare the output with that obtained from previous versions of `lightr` (https://github.com/Bisaloo/lightr/tree/master/tests/testthat/known_output)

> You could improve the package by trying to show (where possible) that the data from your examples corresponds to what one would expect to see.

The main issue here is that other tools to extract this data, including proprietary tools discard a lot of information, as I mention in the README. The only way we know this data is correct is because we know how the measurements were made (i.e. we know the spectrometer model, the integration time, etc.) and how reflectance spectra should look for such an object. But I don't see a convenient way to demonstrate this to the user.

What I can do though is compare the output from:
- [`parse_avantes_binary.R`](https://github.com/Bisaloo/lightr/blob/master/R/parse_avantes_binary.R) with [`parse_avantes_converted.R`](https://github.com/Bisaloo/lightr/blob/master/R/parse_avantes_converted.R)
- [`parse_jdx.R`](https://github.com/Bisaloo/lightr/blob/master/R/parse_jdx.R)/[`parse_procspec.R`](https://github.com/Bisaloo/lightr/blob/master/R/parse_procspec.R) with [`parse_oceanoptics_converted.R`](https://github.com/Bisaloo/lightr/blob/master/R/parse_oceanoptics_converted.R)

for the same data (i.e. before and after export by official proprietary software). This is not straightforward because as mentioned previously, data is processed in an undocumented way while exported (metadata is partially discarded, spectral data is sometimes interpolated with an undocumented method, etc.).

> For example you could extend the examples by plotting or modelling the output data, such that it can be visually inspected. As a dummy user, I would enjoy an intro vignette with a worked example of reading and modelling spectral data, and seeing a pattern appear. This would also help with illustrating how lightr is used in combination with other R packages to perform a complete analysis.

This is definitely a great idea and I will try to add this soon! An realistic example with measurements on one object and the results of the visual models (via `pavo`) next to a picture of this object could be a further way to convince the user that `lightr` at least produces data that is coherent with what we can observe with our naked eye. This would actually rely a bit more on the functionality from `pavo` than that from `lightr` but it doesn't matter.

> For the formats that are also supported by another package, you could consdider adding a unit test to check if both packages return the same data.

Good point. I'll look into this.

> If only proprietary reference software is available, perhaps you could show summary or visual statistics that allow potential users to compare the output with the other software.

This would be addressed in the intro vignette you propose.

> I think more potential users are willing to give the package a try when you show [...] make it convincing that the results correspond to those of other (proprietary) software.

I'm wondering what would be the best way to convey this to the (potential) users. I doubt many users will dig into the `tests` folder. But I don't really see this as a vignette either. Maybe something derived from [`covrpage`](https://github.com/metrumresearchgroup/covrpage)?

----

TODO:
- [ ] Create intro vignette (with hummingbird data?)
- [ ] Check tests from other packages and duplicate them in lightr
- [ ] Compare output of `lightr` for "raw" vs "exported" OceanOptics and Avantes files.
