### 2019-06-12 d0d7fb8
Use tempdir for tests
