### 2019-06-18 9d23997
Add comparison test with photobiologyInOut
