### 2019-06-18 03734ca
Remove mentions of hyperSpec and readJDX After further testing, they do not support reflectance JDX files
