### 2019-07-04 c518cc3
Add emojis for friendlier README
