+++ .travis.yml
@@ -14,6 +14,14 @@ matrix:
       env: DEPLOY=true
 sudo: false
 cache: packages
+addons:
+  apt:
+    sources:
+    - sourceline: 'ppa:cran/imagemagick'
+    packages:
+    - libgmp3-dev
+    - libmagick++-dev
+    - libglu1-mesa-dev
 r_packages:
   - covr
   - goodpractice
