### 2019-07-04 cd41297
Install lightr before deploy
