### 2019-07-23 8b2f51d
Build pkgs on appveyor if necessary See https://twitter.com/mdsumner/status/1153641304808493058
