+++ appveyor.yml
@@ -1,4 +1,8 @@
 # DO NOT CHANGE the "init" and "install" sections below
+environment:
+  global:
+    USE_RTOOLS: true
+    PKGTYPE: both
 
 # Download script file from GitHub
 init:
