### 2019-07-25 #267 @Bisaloo
Hi @karthik, just wondering what's the status here.

I can try to address the remaining concerns from @jeroen in the next couple of weeks, by providing a new vignette which:

- shows that we get the same output with my parsers and the proprietary parsers from Avantes and OceanOptics
- shows that the spectra match what we observe in human vision
- demonstrates how this data can then be used with a vision modelling tool such as pavo

Would be that enough or should I wait for another review?
