### 2019-08-01 #267 @karthik
Hi @Bisaloo,
It would be good for you to address these comments now. I'll be your second reviewer and will provide my review within a week. Once you address any additional issues, we can proceed with the next steps.

PS: I'll be away from email till August 6.
