### 2019-08-08 c02c172
Add test to compare to official export
