+++ tests/testthat/compare_official.R
@@ -1,0 +1,7 @@
+test_that("compare procspec/spectrasuite", {
+
+  specs <- get_spec(test.file("puffin"), ext = "ProcSpec")
+  specs_spectrasuite <- get_spec(test.file("puffin"), ext = "txt")
+
+  expect_equal(specs, specs_spectrasuite, tol = 1e-4)
+})
