### 2019-08-30 f4bdeb5
Make parse_oo_converted work even when closing tag is missing
