### 2019-09-02 e3dd4bc
Add intro vignette
