+++ vignettes/data/puffin/BR_PF26_1.txt
@@ -1,0 +1,3648 @@
+176.36	0.000
+176.58	0.000
+176.80	0.000
+177.02	-254.608
+177.24	-237.047
+177.47	-203.750
+177.69	-252.186
+177.91	-699.422
+178.13	-266.048
+178.35	-242.979
+178.57	-296.392
+178.79	-317.886
+179.01	-333.012
+179.23	-330.916
+179.46	-344.170
+179.68	-349.442
+179.90	-374.453
+180.12	-389.819
+180.34	-427.639
+180.56	-482.416
+180.78	-567.272
+181.00	-602.893
+181.22	-654.416
+181.44	-686.901
+181.66	-798.340
+181.89	-876.103
+182.11	-909.270
+182.33	-1014.726
+182.55	-1335.801
+182.77	-1465.154
+182.99	-1344.412
+183.21	-1410.241
+183.43	-1164.413
+183.65	-1369.488
+183.87	-2479.757
+184.09	-3227.261
+184.31	-5549.103
+184.53	-42077.966
+184.76	8582.474
+184.98	77878.125
+185.20	-7748.734
+185.42	-6466.227
+185.64	-9365.900
+185.86	-5695.581
+186.08	-38587.302
+186.30	-10370.085
+186.52	84972.414
+186.74	4450.626
+186.96	3697.761
+187.18	4291.519
+187.40	7861.688
+187.62	3576.799
+187.84	3326.860
+188.06	3187.682
+188.28	7834.323
+188.51	-18452.381
+188.73	-14598.742
+188.95	-6092.000
+189.17	-3243.262
+189.39	-2443.615
+189.61	-1936.293
+189.83	-2109.302
+190.05	-1943.614
+190.27	-2171.014
+190.49	-2550.851
+190.71	-2943.590
+190.93	-4023.592
+191.15	-3483.614
+191.37	-2773.537
+191.59	-2665.152
+191.81	-2878.483
+192.03	-4077.915
+192.25	-2388.058
+192.47	-2721.183
+192.69	-3127.951
+192.91	-6876.036
+193.13	-5452.000
+193.35	-4027.700
+193.57	-6534.167
+193.79	-18866.935
+194.01	-14740.000
+194.23	-73237.500
+194.45	8912.453
+194.67	20757.018
+194.89	9657.317
+195.11	11425.481
+195.33	5201.762
+195.55	3810.032
+195.77	2979.026
+195.99	3205.108
+196.21	2793.506
+196.43	2278.182
+196.65	2220.130
+196.87	1932.824
+197.09	1696.908
+197.31	1668.407
+197.53	1812.923
+197.75	1707.546
+197.97	1555.712
+198.19	1786.603
+198.41	2330.029
+198.63	2099.058
+198.85	1668.568
+199.07	1724.488
+199.29	1483.792
+199.51	1525.991
+199.73	1482.686
+199.95	1454.221
+200.17	1535.250
+200.39	1421.601
+200.61	1251.703
+200.83	1132.131
+201.05	1244.680
+201.27	1192.508
+201.49	1352.227
+201.71	1353.148
+201.93	1352.899
+202.15	1499.507
+202.37	1464.320
+202.59	1365.183
+202.81	1055.494
+203.03	969.818
+203.25	949.285
+203.47	1087.729
+203.69	1071.261
+203.91	951.155
+204.13	957.154
+204.35	881.613
+204.57	974.343
+204.79	1063.280
+205.01	1060.655
+205.23	1114.311
+205.45	1148.599
+205.67	1105.523
+205.89	1148.133
+206.10	1089.280
+206.32	1004.393
+206.54	925.120
+206.76	798.087
+206.98	778.574
[…]
