+++ vignettes/true_example.Rmd
@@ -55,8 +55,8 @@ We can visually inspect these two file using the dedicated S3 function
 
 ```{r, fig.show='hold'}
 library(pavo)
-plot(raw_files, col = seq(1,6))
-plot(txt_files, col = seq(1,6))
+plot(raw_files, col = seq(1,6), main = "Raw ProSpec files")
+plot(txt_files, col = seq(1,6), main = "Exported txt files")
 ```
 
 Spectra imported by `lightr` and converted by SpectraSuite look pretty similar 
