### 2019-10-15 181a51e
Update colours of specs (one col per patch)
