### 2019-10-17 21a340e
Fix paper.md YML header
