### 2019-10-17 6168e59
Fix bib file and add title to paper
