+++ inst/lightr.bib
@@ -1,5 +1,5 @@
 
-@article{Cuthill1999_plumagereflectanceobjective,
+@article{Cuthill1999,
   langid = {english},
   title = {Plumage Reflectance and the Objective Assessment of Avian Sexual Dichromatism},
   volume = {153},
@@ -11,7 +11,7 @@
   author = {Cuthill, Innes C. and Bennett, Andrew T. D. and Partridge, J. C. and Maier, E. J.}
 }
 
-@article{Bennett1994_sexualselectionmismeasure,
+@article{Bennett1994,
   langid = {english},
   title = {Sexual Selection and the Mismeasure of Color},
   volume = {144},
@@ -23,7 +23,7 @@
   author = {Bennett, Andrew T. D. and Cuthill, Innes C. and Norris, K. J.}
 }
 
-@article{Maia2013_pavopackageanalysis,
+@article{Maia2013,
   langid = {english},
   title = {Pavo: An {{R}} Package for the Analysis, Visualization and Organization of Spectral Data},
   volume = {4},
@@ -37,7 +37,7 @@
   author = {Maia, Rafael and Eliason, Chad M. and Bitton, Pierre-Paul and Doucet, Stéphanie M. and Shawkey, Matthew D.}
 }
 
-@article{White2015_signaldesigncourtship,
+@article{White2015,
   langid = {english},
   title = {Signal Design and Courtship Presentation Coincide for Highly Biased Delivery of an Iridescent Butterfly Mating Signal},
   volume = {69},
@@ -50,7 +50,7 @@
   author = {White, Thomas E. and Zeil, Jochen and Kemp, Darrell J.}
 }
 
-@article{Eaton2005_humanvisionfails,
+@article{Eaton2005,
   langid = {english},
   title = {Human Vision Fails to Distinguish Widespread Sexual Dichromatism among Sexually “Monochromatic” Birds},
   volume = {102},
@@ -65,7 +65,7 @@
   eprint = {16033870}
 }
 
-@article{Maia2018_pavonewtools,
+@article{Maia2018,
   langid = {english},
   title = {Pavo 2.0: New Tools for the Spectral and Spatial Analysis of Colour in {{R}}},
   doi = {10.1101/427658},
