### 2019-10-18 #267 @karthik
Adding to Jeroen's feedback, here is my review. Most comments are minor, and once I see a response, I can move quickly to get this accepted.

# Review of lightr package

The lightr package provides the ability to parse spectrometry data from different file formats and from multiple different scientific instruments. The alternatives are all proprietary software.

### General comments

- In the package down documentation, the two main functions are impossible to test without any example files to read from. When running this in my console, I get:

```
Warning message:
No files found. Try a different value for argument "ext".
```

Which makes sense. At least in the examples, can you point it to some example files that you can include with the package itself?

- Test coverage for the package is great! 🙌

- In the function documentation, when describing specific instruments (e.g. `OceanOptics`, `Avantes`), link to those where possible.

- Unless the package API is set in stone, I would recommend remaining some of the functions. Currently here is what is exported:

```
[1] "convert_tocsv"  "get_metadata"   "get_spec"       "parse_abs"
 [5] "parse_generic"  "parse_jaz"      "parse_jazirrad" "parse_jdx"
 [9] "parse_procspec" "parse_roh"      "parse_trm"      "parse_trt"
[13] "parse_ttt"
```

By adding some sort of standard prefix, e.g. `lr_` it would allow for easy autocompletion in Rstudio and other text editors/IDEs. Functions like `get_metadata` are so generic that it will likely cause namespace conflicts with other packages.

### Minor comments

- Version number still shows as `Version: 0.0.0.9000`. You might want to bump the minor version at least with each major change so it's evident that the package is not super early in development. Leaving the version number at the devtools default does act as a signal.
- The url field has two URLs:
```URL: https://bisaloo.github.io/lightr, https://github.com/Bisaloo/lightr```
Perhaps link to just one? Maybe the package down docs (which then link to the repo itself) so that the field is easy to parse for other purposes. @jeroen do you have other thoughts here on how it might impact the docs parser?
-"If you can't find the best parse fror your specific file"          README.md:119. fror should be for. You can check spelling on the whole package with `spelling::spell_check_package()`
