### 2019-10-19 22ee916
Add links to Avantes & OceanOptics in docs
