### 2019-10-19 2beb31e
Change 'where' arg in README so that it works with git clone
