### 2019-10-19 #267 @Bisaloo
Thank you for your comments @karthik :+1:! please see my changes below:

> ### General comments
>
>   - In the package down documentation, the two main functions are impossible to test without any example files to read from. When running this in my console, I get:
>
> ```
> Warning message:
> No files found. Try a different value for argument "ext".
> ```
>
> Which makes sense. At least in the examples, can you point it to some example files that you can include with the package itself?

https://github.com/Bisaloo/lightr/commit/2beb31ea60f42ad2af8dcd1750767d4d18cf6d03

>  -  Test coverage for the package is great! :raised_hands:
>
>  - In the function documentation, when describing specific instruments (e.g. OceanOptics, Avantes), link to those where possible.

https://github.com/Bisaloo/lightr/commit/22ee9164afde74348df2625101d6cc0d884e3ee1

 >   Unless the package API is set in stone, I would recommend remaining some of the functions. Currently here is what is exported:
>
> ```
> [1] "convert_tocsv"  "get_metadata"   "get_spec"       "parse_abs"
> [5] "parse_generic"  "parse_jaz"      "parse_jazirrad" "parse_jdx"
> [9] "parse_procspec" "parse_roh"      "parse_trm"      "parse_trt"
> [13] "parse_ttt"
> ```
>
> By adding some sort of standard prefix, e.g. `lr_` it would allow for easy autocompletion in Rstudio and other text editors/IDEs. Functions like `get_metadata` are so generic that it will likely cause namespace conflicts with other packages.

Would you do this for all exported functions or just the three main ones which are the most likely to have name conflicts (`get_spec()`, `get_metadata()` and `convert_tocsv()`)?

> ### Minor comments
>
>  - Version number still shows as Version: 0.0.0.9000. You might want to bump the minor version at least with each major change so it's evident that the package is not super early in development. Leaving the version number at the devtools default does act as a signal.

https://github.com/Bisaloo/lightr/commit/b9159922db33c7b19c31de4c7155f2ae9606d82a

> -    The url field has two URLs:
    `URL: https://bisaloo.github.io/lightr, https://github.com/Bisaloo/lightr`
    Perhaps link to just one? Maybe the package down docs (which then link to the repo itself) so that the field is easy to parse for other purposes. @jeroen do you have other thoughts here on how it might impact the docs parser?

Actually, the recommended practice for pkgdown is to have both URL. The `github.com` URL is required to auto-link to the github repo and fill out the bug report field (in the sidebar and the github logo in the top navbar). The `github.io` is required for cross-linking from other packages (https://pkgdown.r-lib.org/articles/linking.html#across-packages).

>   - "If you can't find the best parse fror your specific file" README.md:119. fror should be for. You can check spelling on the whole package with spelling::spell_check_package()

https://github.com/Bisaloo/lightr/commit/515d193373b35d1faddc117c6606a97ca7a32c74
