### 2019-10-19 #267 @karthik
> Would you do this for all exported functions or just the three main ones which are the most likely to have name conflicts (get_spec(), get_metadata() and convert_tocsv())?

Generally for all exported functions unless you have a strong reason not to.
