### 2019-10-20 a0ed298
Add lr_ prefix to all exported functions
