+++ tests/testthat/compare_photobiologyInOut.R
@@ -8,7 +8,7 @@ test_that("pio_lightr_jaz", {
   out_pio[,2] <- out_pio[,2] * 100
   colnames(out_pio) <- NULL
 
-  out_lightr <- parse_jazirrad(file)[[1]]
+  out_lightr <- lr_parse_jazirrad(file)[[1]]
   colnames(out_lightr) <- NULL
 
   expect_equivalent(out_pio, out_lightr[, c(1,5)])
