### 2019-10-20 #267 @Bisaloo
Ok, thanks! Done in https://github.com/Bisaloo/lightr/commit/a0ed29867f3bc8d36b92f5b22f58c32de936ccab.
