### 2019-10-21 #267 @karthik
To run the examples in your README, you'll need a slight change to locate the `inst/testdata/` files

```
lr_get_metadata(system.file("testdata/procspec_files" , package = "lightr"),
                ext = "ProcSpec")
```

Same for the example below.
