### 2019-10-21 62d59b7
Edit 'where' arg in README to work with installed pkg
