### 2019-10-21 #267 @Bisaloo
Right, it makes more sense since users have to install the package to run this anyways :+1:. It's done!
