### 2019-10-21 41c6b91
Add ropensci footer
