### 2019-10-21 bd8bb47
Update all links to ropensci/lightr
