+++ README.md
@@ -1,9 +1,9 @@
 # `lightr`: import spectral data in R
 
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![Travis build status](https://travis-ci.org/Bisaloo/lightr.svg?branch=master)](https://travis-ci.org/Bisaloo/lightr)
-[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/Bisaloo/lightr?branch=master&svg=true)](https://ci.appveyor.com/project/Bisaloo/lightr)
-[![Coverage status](https://codecov.io/gh/Bisaloo/lightr/branch/master/graph/badge.svg)](https://codecov.io/github/Bisaloo/lightr?branch=master)
+[![Travis build status](https://travis-ci.org/ropensci/lightr.svg?branch=master)](https://travis-ci.org/ropensci/lightr)
+[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/ropensci/lightr?branch=master&svg=true)](https://ci.appveyor.com/project/ropensci/lightr)
+[![Coverage status](https://codecov.io/gh/ropensci/lightr/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/lightr?branch=master)
 [![Under review at rOpenSci](https://badges.ropensci.org/267_status.svg)](https://github.com/ropensci/software-review/issues/267)
 
 There is no standard file format for spectrometry data and different scientific
@@ -26,7 +26,7 @@ This package is not yet published on CRAN and must be installed via GitHub:
 
 ```r
 # install.packages("remotes")
-remotes::install_github("bisaloo/lightr")
+remotes::install_github("bropenscilightr")
 ```
 
 ## 💻 Usage
@@ -123,7 +123,7 @@ flexible and general algorithm that manages to extract data from most files.
 
 If you can't find the best parser for your specific file or if you believe you
 are using an unsupported format, please 
-[open an issue](https://github.com/Bisaloo/lightr/issues) or send me an email. 
+[open an issue](https://github.com/Biropensciightr/issues) or send me an email. 
 
 ## 🌐 Similar projects
 
