+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://bisaloo.github.io/lightr
+url: https://docs.ropensci.org/lightr
 
 development:
   mode: auto
