### 2019-10-21 #267 @karthik
@Bisaloo Once you transfer, the documentation will also be auto generated at docs.ropensci.org/lightr

You'll need to update that URL in your DESCRIPTION (and the top of your repo) as well after the transfer.
