### 2019-10-21 6f1800c
Remove auto-deploy
