### 2019-10-21 #267 @Bisaloo
Should I go ahead and submit this to JOSS once I add a DOI?
