### 2019-10-21 #267 @karthik
You can add a DOI after the JOSS submission is accepted. So go ahead and submit there. In the meantime, you can compile your own paper here: http://whedon.theoj.org/ and fix any mistakes or issues such as broken references.

Also reading over your current paper, you provide no examples of a scientific application. That will be necessary for this software to get accepted to JOSS.
