### 2019-10-21 #267 @Bisaloo
> Also reading over your current paper, you provide no examples of a scientific application. That will be necessary for this software to get accepted to JOSS.

Okay, thanks for the advice! I'll do this tomorrow then.
