### 2019-10-25 #267 @karthik
I think it would be useful to use a test dataset + lightr to show what can be done as a small example.  [Here](https://joss.theoj.org/papers/10.21105/joss.01593) is an example. If you submit as is, that's the first question an editor will as you about.
